Simplify tools/cargo_package.py
