+++ core/BUILD.gn
@@ -68,50 +68,3 @@ rust_test("deno_core_http_bench_test") {
   ]
   extern = http_bench_extern
 }
-
-if (is_linux) {
-  llvm_target = "x86_64-unknown-linux-gnu"
-  static_lib_suffix = ".a"
-} else if (is_mac) {
-  llvm_target = "x86_64-apple-darwin"
-  static_lib_suffix = ".a"
-} else if (is_win) {
-  llvm_target = "x86_64-pc-windows-msvc"
-  static_lib_suffix = ".lib"
-} else {
-  assert(false)
-}
-
-deno_core_target_dir = "$root_out_dir/cargo_pkg/deno_core_target"
-
-copy("deno_core_target_static_lib") {
-  lib_name = "libdeno$static_lib_suffix"
-  sources = [
-    "$root_out_dir/obj/core/libdeno/$lib_name",
-  ]
-  outputs = [
-    "$deno_core_target_dir/lib/$lib_name",
-  ]
-  deps = [
-    "libdeno:libdeno_static_lib",
-  ]
-}
-
-action("deno_core_target") {
-  script = "//tools/cargo_pkg_target.py"
-  args = [
-    "--root",
-    rebase_path("//", root_build_dir),
-    "--out",
-    rebase_path(deno_core_target_dir, root_build_dir),
-    "--target",
-    llvm_target,
-  ]
-  deps = [
-    ":deno_core_target_static_lib",
-    "libdeno:libdeno_static_lib",
-  ]
-  outputs = [
-    "$deno_core_target_dir/Cargo.toml",
-  ]
-}
